+++ Day5/ConversionRange.cs
@@ -3,12 +3,12 @@
     internal class ConversionRange
     {
         // Both bounds are inclusive.
-        private readonly int _lowerSourceBound;
-        private readonly int _upperSourceBound;
+        private readonly Int64 _lowerSourceBound;
+        private readonly Int64 _upperSourceBound;
 
-        private readonly int _conversion;
+        private readonly Int64 _conversion;
 
-        public ConversionRange(int pLowerSourceBound, int pUpperSourceBound, int pConversion)
+        public ConversionRange(Int64 pLowerSourceBound, Int64 pUpperSourceBound, Int64 pConversion)
         {
             _lowerSourceBound = pLowerSourceBound;
             _upperSourceBound = pUpperSourceBound;
@@ -16,12 +16,12 @@
             _conversion = pConversion;
         }
 
-        private bool IsSourceIncluded(int pSource)
+        private bool IsSourceIncluded(Int64 pSource)
         {
             return pSource >= _lowerSourceBound && pSource <= _upperSourceBound;
         }
 
-        public bool Convert(int pSource, out int pDestination)
+        public bool Convert(Int64 pSource, out Int64 pDestination)
         {
             bool isSourceWithinRange = IsSourceIncluded(pSource);
             pDestination = isSourceWithinRange ? pSource + _conversion : pSource;
